+++ src/api/hooks/index.ts
@@ -1,11 +1,9 @@
-export * from './use-assign-quest'
-export * from './use-repeat-quest'
 export * from './use-major-interest'
-export * from './use-quest-setting'
 
-export * from './interest'
 export * from './history'
+export * from './interest'
 export * from './member'
+export * from './quest'
 export * from './streak'
 
 export * from './api-mutate'
